+++ android/src/main/java/com/sudoplz/rninappupdates/SpReactNativeInAppUpdatesPackage.java
@@ -1,28 +1,45 @@
 package com.sudoplz.rninappupdates;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
-import com.facebook.react.ReactPackage;
+import com.facebook.react.TurboReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.uimanager.ViewManager;
+import com.facebook.react.module.model.ReactModuleInfo;
+import com.facebook.react.module.model.ReactModuleInfoProvider;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import java.util.HashMap;
+import java.util.Map;
 
-public class SpReactNativeInAppUpdatesPackage implements ReactPackage {
-    @NonNull
+public class SpReactNativeInAppUpdatesPackage extends TurboReactPackage {
+    @Nullable
     @Override
-    public List<NativeModule> createNativeModules(@NonNull ReactApplicationContext reactContext) {
-        List<NativeModule> modules = new ArrayList<>();
-        modules.add(new SpReactNativeInAppUpdatesModule(reactContext));
-        return modules;
+    public NativeModule getModule(@NonNull String name, @NonNull ReactApplicationContext reactContext) {
+        if (name.equals(SpReactNativeInAppUpdatesModule.NAME)) {
+            return new SpReactNativeInAppUpdatesModule(reactContext);
+        }
+        return null;
     }
 
     @NonNull
     @Override
-    public List<ViewManager> createViewManagers(@NonNull ReactApplicationContext reactContext) {
-        return Collections.emptyList();
+    public ReactModuleInfoProvider getReactModuleInfoProvider() {
+        return () -> {
+            final Map<String, ReactModuleInfo> moduleInfos = new HashMap<>();
+            moduleInfos.put(
+                SpReactNativeInAppUpdatesModule.NAME,
+                new ReactModuleInfo(
+                    SpReactNativeInAppUpdatesModule.NAME,
+                    SpReactNativeInAppUpdatesModule.class.getName(),
+                    false, // canOverrideExistingModule
+                    false, // needsEagerInit
+                    true,  // hasConstants
+                    false, // isCxxModule
+                    true   // isTurboModule
+                )
+            );
+            return moduleInfos;
+        };
     }
 }